Extract viewmodel to its own file
